x/axinomixrunner/types/message_end_race_test.go
package types

import (
	"testing"

	"axinomix-runner/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgEndRace_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgEndRace
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgEndRace{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgEndRace{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
